checked slab reader

+++ src/H5util.hpp
@@ -16,8 +16,8 @@ std::vector<double> read_doubles(const char * filename,
 
 std::vector<double> read_doubles_slab(const char * filename,
 				      const char * dsetname,
-				      const int & start,
-				      const int & len);
+				      const size_t & start,
+				      const size_t & len);
 
 void write_strings( const std::vector<std::string> & data,
 			 const char * dsetname,
